\name{nwsClose}
\docType{methods}
\alias{nwsClose}
\alias{nwsClose,netWorkSpace-method}
\title{netWorkSpace Class Method}
\description{
Close connection of a shared netWorkSpace to the netWorkSpace server.
}

\usage{
nwsClose(.Object)
}

\arguments{
\item{.Object}{a netWorkSpace class object}
}

\examples{
\dontrun{
ws = new('netWorkSpace', 'nws example')
# do some works
# ...
nwsClose(ws)
}
}

\keyword{methods}
