\name{nwsPkgInfo}
\alias{nwsPkgInfo}
\title{NetWorkSpaces Package Information}
\description{
Reports package name and version information.
}

\usage{
nwsPkgInfo()
}

\value{
A character vector with with fields
\item{name}{The package name.}
\item{version}{The version.}
}

\details{
\code{nwsPkgInfo} returns details of the NetWorkSpaces package.
The name can be either \code{"nws"} or \code{"nwsPro"}.
}

\examples{
nwsPkgInfo()
}

\keyword{utilities}
