\name{nwsWsName}
\docType{methods}
\alias{nwsWsName}
\alias{nwsWsName,netWorkSpace-method}
\title{netWorkSpace Class Method}
\description{
Return name of a netWorkSpace.
}

\usage{
nwsWsName(.Object)
}

\arguments{
\item{.Object}{a netWorkSpace class object}
}

\examples{
\dontrun{
ws = new('netWorkSpace', 'nws example')
nwsWsName(ws)
}
}
\keyword{methods}
