\name{lsfcmd}
\alias{lsfcmd}
\title{sleigh Auxiliary Function}
\description{
This function is used to indicate LSF will be used to launch sleigh workers. 
}
\seealso{\code{\link{sleigh}}}

\details{
This function is only meant to be passed into sleigh constructor.
}

\examples{
\dontrun{
s = sleigh(launch=lsfcmd)
}
}

\keyword{methods}
\keyword{internal}