\name{nwsClose}
\alias{nwsClose}
\alias{nwsClose,netWorkSpace-method}
\title{Close connection to shared netWorkSpace server}
\description{
Close connection to shared netWorkSpace server.
}

\usage{
nwsClose(.Object)
}

\arguments{
  \item{.Object}{A netWorkSpace object}
}

\examples{
\dontrun{
ws = new("netWorkSpace", "my Space")
# ... (use ws)
nwsClose(ws)
}
}

\keyword{methods}
\keyword{programming}
