\name{nwsWsName}
\alias{nwsWsName}
\alias{nwsWsName,netWorkSpace-method}
\title{returns name of a netWorkSpace}

\description{
returns name of a netWorkSpace.
}

\usage{
nwsWsName(.Object)
}

\arguments{
   \item{.Object}{A netWorkSpace object}
}

\details{
Returns the name of an object nwtWorkSpace as a string,
}
\keyword{methods}
\keyword{programming}
