\name{stopSleigh}
\alias{stopSleigh}
\title{Shutdown workers and remove netWorkSpace}
\description{
Shutdown workers and remove netWorkSpace.
}

\usage{
stopSleigh(.Object)
}

\arguments{
\item{.Object}{a sleigh object}
}

\examples{
\dontrun{
s = sleigh()
stopSleigh(s)
}
}
\keyword{methods}
\keyword{programming}

